refactor: remove unused old seed matching code
Old `seed_match.rs` removed

New `seed_match2.rs` -> `seed_match.rs`
