+++ packages/nextclade/src/align/mod.rs
@@ -10,4 +10,3 @@ pub mod score_matrix_aa;
 pub mod score_matrix_nuc;
 pub mod seed_alignment;
 pub mod seed_match;
-pub mod seed_match2;
